// src/app/models/host/planning-command-host-send-type.enum.ts
export enum PlanningCommandHostSendType {
    startSession = "START_SESSION",
    addTicket = "ADD_TICKET",
    skipVote = "SKIP_VOTE",
    removeParticipant = "REMOVE_PARTICIPANT",
    endSession = "END_SESSION",
    finishVoting = "FINISH_VOTING",
    revote = "REVOTE",
    reconnect = "RECONNECT"
}